requirements.txt
fastapi
uvicorn
pdfplumber
sentence-transformers
faiss-cpu
pydantic
python-dotenv